# docs/swagger.yaml
basePath: /api/v1
definitions:
  handlers.Error:
    properties:
      error:
        type: string
    type: object
  sessionhandler.deleteSessionBody:
    properties:
      token:
        type: string
    required:
    - token
    type: object
  sessionhandler.loginRequestBody:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  userhandler.userBody:
    properties:
      display_name:
        type: string
      password:
        minLength: 8
        type: string
      username:
        type: string
    required:
    - display_name
    - password
    - username
    type: object
  userhandler.userGetResponse:
    properties:
      display_name:
        type: string
      username:
        type: string
    type: object
  userhandler.userPatchBody:
    properties:
      displayName:
        type: string
    type: object
info:
  contact: {}
  description: SnackLog API for managing users and sessions
  title: Authentication API for SnackLog
  version: "1.0"
paths:
  /auth/session:
    delete:
      consumes:
      - application/json
      description: Revokes a user session
      parameters:
      - description: Token to revoke
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/sessionhandler.deleteSessionBody'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.Error'
      security:
      - BearerAuth: []
      summary: Logout
      tags:
      - session
    get:
      description: Retrieves session information by its token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
      security:
      - BearerAuth: []
      summary: Get session info
      tags:
      - session
    post:
      consumes:
      - application/json
      description: Authenticates a user and returns a token
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/sessionhandler.loginRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.Error'
      summary: Login
      tags:
      - session
  /auth/user:
    delete:
      description: Deletes the authenticated user account
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.Error'
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - user
    get:
      description: Retrieves the profile of the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/userhandler.userGetResponse'
        "401":
          description: Unauthorized
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.Error'
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - user
    patch:
      consumes:
      - application/json
      description: Updates the profile of the authenticated user
      parameters:
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/userhandler.userPatchBody'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Error'
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.Error'
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - user
    post:
      consumes:
      - application/json
      description: Creates a new user account
      parameters:
      - description: User registration details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/userhandler.userBody'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.Error'
      summary: Register
      tags:
      - user
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"